feat: enable dynamic checkouts with also a US-Shop
