+++ src/config.js
@@ -4,4 +4,6 @@ module.exports = {
   IVY_CDN_URL: process.env.IVY_CDN_URL,
   IVY_API_URL: process.env.IVY_API_URL,
   IVY_WEBHOOK_SIGNING_SECRET: process.env.IVY_WEBHOOK_SIGNING_SECRET,
+  US_IVY_API_KEY: process.env.US_IVY_API_KEY,
+  US_IVY_WEBHOOK_SIGNING_SECRET: process.env.US_IVY_WEBHOOK_SIGNING_SECRET,
 }
