+++ src/server.js
@@ -8,6 +8,8 @@ const cors = require('cors')
 const webhookRoutes = require('./routes/webhooks')
 const routes = require('./routes')
 const config = require('./config')
+const callbacks = require('./routes/callbacks')
+const validateRequest = require('./middleware/validate')
 
 const server = express()
 server.set('view engine', 'pug')
@@ -22,6 +24,10 @@ server.use(express.static(path.join(__dirname, 'assets')))
 
 server.use('/webhooks', webhookRoutes)
 server.use(routes)
+
+server.use('/callback', validateRequest(config.IVY_WEBHOOK_SIGNING_SECRET), callbacks)
+//Stuff for us sample shop
+server.use('/us/callback', validateRequest(config.US_IVY_WEBHOOK_SIGNING_SECRET), callbacks)
 
 module.exports = async () => {
   server.listen(config.PORT, () => {
